dht_chord

Module api_communication

Provides a server socket for API communication

The API-communications module is responsible for communication with other modules. Incoming requests are dispatched to a worker thread pool, that will process each request independently and asynchronously.

Answer to messages received are not guaranteed to be sent in the same order as the requests were received.

As per specification, the following requests are accepted and processed:

DHT PUT

DHT GET

DHT SUCCESS

DHT FAILURE

We introduced an additional API Message, API_DHT_SHUTDOWN which allows us to shut down a node gracefully through the API. The package has a fixed size and does not contain any other information except the header:

+--------+--------+--------+--------+
|        4        |  DHT SHUTDOWN   |
+--------+--------+--------+--------+

All ApiPacket s we receive, are first parsed by their ApiPacketHeader. Depending on the header, we then parse the rest of the message into the corresponding ApiPacketMessage.

Structs

ApiPacket

Internal representation of all packages received on the API socket

ApiPacketHeader

Header of an ApiPacket

DhtGet

Internal representation of a DHT_GET request

DhtGetFailure

Internal representation of a DHT_FAILURE response

DhtGetSuccess

Internal representation of a DHT_SUCCESS response

DhtPut

Internal representation of a DHT_PUT request

Enums

ApiPacketMessage

Content of an ApiPacket

Constants

API_DHT_FAILURE

Answer to a failed API_DHT_GET request

API_DHT_GET

Requests the DHT to retrieve a value

API_DHT_PUT

Requests the DHT to store a value

API_DHT_SHUTDOWN

Requests our DHT node to shutdown

API_DHT_SUCCESS

Answer to a successful API_DHT_GET request

Functions

hash_key_bytes 🔒

process_api_get_request 🔒

process_api_put_request 🔒

start_api_server 🔒