dht_chord::chord::peer_messages

Enum PeerMessage

pub enum PeerMessage {

    GetNode(u64),
    GetNodeResponse(ChordPeer),
    GetValue(u64),
    GetValueResponse(Option<Vec<u8>>),
    InsertValue(u64, Vec<u8>, Duration),
    SplitRequest(ChordPeer),
    SplitResponse(SplitResponse),
    GetPredecessor,
    GetPredecessorResponse(ChordPeer),
    SetPredecessor(ChordPeer),
    SetSuccessor(ChordPeer),
    ProofOfWorkChallenge(ProofOfWorkChallenge),
    ProofOfWorkResponse(ProofOfWorkResponse),
    CloseConnection,

}

All communication messages sent between peers

Whenever we transmit a message between nodes, we transmit this enum and match on it on the receiving end.

Variants

GetNode(u64)

Get node responsible for key

This is the “backbone message” of our Chord-DHT. It can be sent to any node in the network, which will then respond with the IP address of the node responsible for the given key. As we can use this message to lookup both keys and nodes, it is the only message type that needs to propagate recursively through the network.

GetNodeResponse(ChordPeer)

Response to PeerMessage::GetNode

GetValue(u64)

Get value from responsible node

Retrieves content from one node and one node only. It must be sent to the node that is responsible for storing that specific id and will not propagate through the network.

GetValueResponse(Option<Vec<u8>>)

Response to PeerMessage::GetValue

InsertValue(u64, Vec<u8>, Duration)

Insert value into responsible node

Inserts content into the network; counterpart to GetValue(id). It must only be sent to the node responsible for the key that the value is to be stored under.

SplitRequest(ChordPeer)

Requests a node to split

The target node performs some sanity checks, ensuring that there is no other node between the new node and itself.

SplitResponse(SplitResponse)

Response to PeerMessage::SplitRequest

If a node is between the new node and the target node, it returns the address of the node between them. Otherwise it returns a list of values the new node is now responsible for.

GetPredecessor

Retrieves the predecessor of a node

GetPredecessorResponse(ChordPeer)

Response to PeerMessage::GetPredecessor

SetPredecessor(ChordPeer)

Sets the predecessor of a node

SetSuccessor(ChordPeer)

Sets the successor of a node

ProofOfWorkChallenge(ProofOfWorkChallenge)

Requests a node to solve a ProofOfWorkChallenge

ProofOfWorkResponse(ProofOfWorkResponse)

Response to PeerMessage::ProofOfWorkChallenge

CloseConnection

Signals intent to gracefully close the connection; acting as EOF