dht_chord::chord

Struct Chord

pub struct Chord {
    state: Arc<ChordState>,
}

Distributed Hash Table

This struct provides methods necessary to interact with the DHT.

As the DHT is heavily multithreaded, its internal state is held in an atomic reference counter (Arc).

Fields
state: Arc<ChordState>
Implementations

impl Chord

pub async fn new( initial_peer: Option<SocketAddr>, server_address: SocketAddr, default_storage_duration: Duration, max_storage_duration: Duration, ) -> Self

Construct new instance of chord.

If no initial peer is provided, it is assumed that this node is the only node present.

Otherwise the initial peer is contacted to determine the successor and predecessor of this node. They are then contacted to transfer responsibility of the entries this node is responsible for.

The finger table will be initialized with all entries pointing to our successor until the first run of housekeeping.

pub async fn start_server_socket( &self, cancellation_token: CancellationToken, ) -> JoinHandle<()>

Starts the server socket to listen for incoming peer connections The cancellation token can be used to gracefully stop the server socket

pub async fn start_housekeeping_thread( &self, cancellation_token: &CancellationToken, ) -> JoinHandle<()>

Method to start the housekeeping thread

This thread will stop when the cancellation token is cancelled

async fn housekeeping(&self, cancellation_token: CancellationToken)

Performs the housekeeping which performs some periodic tasks

Cleanup of expired values in ChordState::personal_storage and ChordState::node_storage

Refresh keys and their replications on other nodes

Stabilize to account for churn

Update finger table to optimize routing

pub async fn insert( &self, key: u64, value: Vec<u8>, ttl: Duration, replication_amount: u8, ) -> Result<()>

Inserts a key value pair into the network

If specified, multiple replications of the value will be inserted. All inserted values are periodically rebroadcasted to counter churn.

async fn internal_insert( &self, key: u64, value: Vec<u8>, ttl: Duration, ) -> Result<()>

Inserts an entry into our local node_storage, alongside its expiry time.

pub async fn get(&self, key: u64) -> Option<Vec<u8>>

Queries the network for a given key and returns its corresponding value.

If the value can not be found on the first try, we also query for replicated backup keys.

fn am_responsible_for_key(&self, key: u64) -> bool

Returns whether this node is responsible for a given key.

This is the case if the key lies between us (inclusive) and our predecessor (exclusive) or no predecessor is currently known, in which case this method returns true.

async fn get_responsible_node(&self, key: u64) -> Result<ChordPeer>

Returns the ChordPeer responsible for a given key.

This method attempts to recursively find the node responsible for a given key by asking the most closely responsible node in our finger table to find it for us.

pub(crate) async fn stabilize(&self) -> Result<()>

Stabilized this node

We contact our predecessor:

If the predecessor is reachable, we check that this node and the predecessor are in agreement about no nodes existing between them. If there are any, we set them as our predecessor

If the predecessor is not reachable, we will iteratively step through our predecessor list in ChordState::predecessors, until an active node is found. We then send this predecessor the message that we are his successor.

Once the predecessor has been stabilized, three more predecessors are recursively acquired to guard against churn.

We contact our successor (the first node in our ChordState::finger_table):

If the successor is reachable, we again ensure that there are no nodes between us or update the finger table accordingly.

If the successor is not reachable, the next peer in the finger table is contacted, until we either find a working node or reach the end of the finger table.

async fn stabilize_predecessor(&self, predecessor_index: usize) -> Result<()>

Attempts to contact the predecessor at the given index. If this predecessor is not reachable, or any other issue occurs, an error is returned. If this connection is successful the predecessor list is modified accordingly.

async fn stabilize_successor( &self, possible_successor: ChordPeer, ) -> Result<ChordPeer>

Attempts to contact the given successor. If the successor is not reachable, or any other issue occurs, an error is returned. If the successor is reachable, this method will iteratively find any reachable predecessor of the successor, and return the last reachable successor before this node.

pub(crate) async fn fix_fingers(&self) -> Result<()>

Iterates through all finger table entries and asks the previous entry about the peer in the next entry.

This ensures that all finger table entries point to the successor of 2^index, where index is the index in the finger table.

fn as_chord_peer(&self) -> ChordPeer

Returns a ChordPeer representing this node; useful for peer communication.

async fn ask_for_predecessor(&self, node_to_ask: ChordPeer) -> Result<ChordPeer>

Returns a result with the predecessor of node_to_ask. If the provided peer is not reachable an error is returned

Arguments

node_to_ask - the node to ask for its predecessor

async fn ask_for_successor(&self, node_to_ask: ChordPeer) -> Result<ChordPeer>

Asks a ChordPeer for its successor

fn get_predecessor(&self) -> ChordPeer

Returns the predecessor of this node. If we do not have a successor, we return ourselves as ChordPeer.

async fn accept_peer_connection(&self, stream: TcpStream) -> Result<()>

Handle incoming request from a connecting peer.

fn id_at_finger_index(&self, index: usize) -> u64

Returns the id which is 2^index after this node

fn assert_finger_table_invariants_correct(&self)

Asserts that we do not contain ourselves in the finger table.

If a node would contain itself, there is a possibility of endless routing loops.