Struct dht_chord::chord::ChordState

struct ChordState {
    node_id: u64,
    address: SocketAddr,
    finger_table: Vec<RwLock<ChordPeer>>,
    predecessors: RwLock<Vec<ChordPeer>>,
    default_storage_duration: Duration,
    max_storage_duration: Duration,
    default_replication_amount: u8,
    personal_storage: DashMap<u64, (Vec<u8>, DateTime<Utc>, u8)>,
    node_storage: DashMap<u64, (Vec<u8>, DateTime<Utc>)>,
}

All data necessary for DHT operation

Fields
node_id: u64

Our own Node-ID.

address: SocketAddr

Our outward facing address.

Note that as it is used to calculate the node id, it must be the address we are reachable under to other nodes.

finger_table: Vec<RwLock<ChordPeer>>

The characteristic Chord jump table. We maintain 64 entries with exponentially growing distance between them. As we are expected to do maintenance and handle churn - even under heavy load, entries are lockable individually. Small inconsistencies are tolerable, as they only affect routing efficiency.

predecessors: RwLock<Vec<ChordPeer>>

List of predecessors do determine key responsibility and provide limited fault tolerance in case of churn.

default_storage_duration: Duration

Default storage duration of entries, if not specified otherwise.

max_storage_duration: Duration

The maximum duration for which we keep an item we have been tasked to store by the network in our node_storage. Does not apply to personal_storage.

default_replication_amount: u8

The default amount of replications which are done for each entry. Replications are performed by increasing the key the entry is stored under by 1 and re-hashing the result.

personal_storage: DashMap<u64, (Vec<u8>, DateTime<Utc>, u8)>

The personal storage keeps track of entries we have been asked to store by the local user. It is used as replication source by the housekeeping thread, ensuring its contents are continuously re-spread across the network. It maps a key to a tuple of the value, the expiration time and the replication amount. It is also used as cache for fast local lookups.

node_storage: DashMap<u64, (Vec<u8>, DateTime<Utc>)>

The node storage keeps track of entries that have been assigned to the local DHT node by the network. Once the stored DateTime is hit, the entry expires and is removed by the housekeeping thread.