Module dht_chord::chord

Implementation of a distributed hash table based on Chord

It can operate independently from the API module and is usable as a stand-alone crate.

Features:

Key-Value storage

Distributed, if more than one node available (but fully functional with one node only)

Built-in replication

IPv4 and IPv6 support

Automatic node discovery

Stabilization if nodes leave or join

Housekeeping thread, to remove expired entries and refresh keys we have been tasked to store

Completely asynchronous and multi-threaded

Requests from the API and from other nodes are processed and answered concurrently

Free of race conditions due to Rusts ownership model

Performance optimized implementation, capable of establishing a fully connected network with 2000 nodes running on a single machine in under 10 seconds (without proof-of-work enabled)

Architecture:

Our architecture separates the Chord module from the API communication

The API only makes the features of the Chord module accessible via network communication

Threading:

We make heavy use of multithreading/processing:

Tokio (green) threads for all asynchronous workloads

Every connecting API and P2P client gets its own thread

Housekeeping is performed in a background thread without blocking other operations

Peer-to-peer communication:

We are using channels for inter-node communication

This allows us to serialize/deserialize entire structs typesafe and integrity checked

Protocol details are specified in the peer_messages module

Inter-node messages are specified in PeerMessage

Nodes exiting unexpectedly or sending invalid messages do not crash other nodes

Errors are detected (and logged if desired) and the connection to the misbehaving node is closed

Non-responding nodes are detected by the housekeeping thread and removed from the overlay

This gives us robustness against non-byzantine faults

Nodes that appear to perform correctly, yet send well-formed but disruptive messages can not be detected

Security features:

SHA-3-512 proof of work challenges with adjustable difficulty for requests

Does not prevent byzantine nodes from splitting the network or eclipsing nodes, but prevents greedy nodes from abusing the storage system

Difficulty is independently adjustable for each request type

Currently we maintain two difficulty settings, one for get requests and one for (potentially disruptive) write/storage requests

Addresses of nodes assigned based on hash of IP + port

This could be easily adjusted to only hash IPs, providing limited sybil defense

In case of IPv6, we could only hash a masked version of the IP

We have currently not implemented this, as it would interfere with development and testing

Limited defence against nodes refusing to store values or disconnecting

Our housekeeping thread continuously refreshes values that have been stored in the DHT upon our own request

Security discussion:

Byzantine fault tolerance is extremely difficult to achieve in a distributed system

We initially tried to implement a byzantine fault tolerance according to this paper, but came to the conclusion that it was not feasible to implement in a reasonable time frame

Byzantine fault tolerance would require multiple complex submodules to be implemented, such as

Secure multiparty computation

Secure network size estimation

A split of the overlay network into multiple swarms

We were not able to find a sufficiently advanced crate for secure multiparty computation and a (byzantine fault tolerant) network size estimation would carry the workload of an entire additional module

This is why we ultimately decided to only implement proof-of-work as a defence against greedy nodes

We understand that our system can not defend against a malicious attacker deliberately providing nodes with false information

Implementing such a system would not only be extremely difficult, but also extremely resource intensive and inefficient, as a significant amount of nodes would need to reach consensus on every single request

We do however note, that a few of our design choices should make attacks more difficult:

Choosing IDs based on their IP address makes it harder for nodes to choose their position in the network

This makes it more difficult to eclipse nodes

The finger table also offers some resistance against eclipse attacks:

Nodes regularly check in with all their fingers, which makes it difficult to eclipse them fully

To prevent race conditions, there are checks if overlay changes like SetPredecessor are valid, or if there are other nodes between the current node and the new predecessor

This makes it more difficult for an attacker to split the network or eclipse nodes, as nodes will not accept an incorrectly placed node any new node as their predecessor

Our stabilization method regularly checks the health of peers in the neighborhood and fixes the overlay if necessary

An attacker needs to operate their peers continuously, as they would otherwise be removed

Overwriting values is possible, but expensive due to proof-of-work

DoS Attacks, such as content pollution and index poisoning, are hardened against by increased proof-of-work difficulty for write requests

Future Work:

We have some ideas/suggestions on how to further improve our implementation:

Improved sybil defence:

Currently we hash the IP and port of a node to determine its ID

This could be easily adjusted to only hash IPs, making it harder to choose a specific node position for an attacker, as this would require them to have control over a large number of IP addresses

New nodes should be treated differently, i.e. not as trustworthy until they stayed some time in the network

We could introduce active scanning measures to track whether nodes are truly active and responsive

Misbehaviour defence:

The most efficient way to “cheaply” detect misbehaving nodes would probably be an out-of-band reporting system and/or a scanning authority that secretly scans for misbehaving nodes and blacklists them, similarly to how The Tor Project detects and blacklists misbehaving relays

This would however, go against the decentralization aspect of our system

Better Stabilize:

Stabilize in its current form relies on each node to individually realize that a peer has disconnected from the network

This sometimes incorrectly invalidates SetPredecessor and SetSuccessor requests, as they are denied on the grounds that the receiving node does not yet know, that its current successor/predecessor no longer exists

Modules

peer_messages

Communication between peers

Macros

connect_to_peer 🔒

Structs

Chord

Distributed Hash Table

ChordState 🔒

All data necessary for DHT operation

Functions

calculate_hash 🔒

Calculate hash for ID-mapping

is_between_on_ring 🔒

Check if a key is between two keys/nodes on the ring

require_proof_of_work 🔒

Sends a ProofOfWorkChallenge to the connected peer and waits for the corresponding peer_messages::ProofOfWorkResponse.

solve_proof_of_work 🔒

Solves a ProofOfWorkChallenge and sends the corresponding peer_messages::ProofOfWorkResponse.